ReviewsandRatingsMicroservice/requirements.txt
Flask==2.2.2
pandas==2.0.1
pymongo==4.3.3
PyJWT~=2.6.0
Flask-Cors==3.0.10
Flask-RESTPlus==0.13.0
flask-restx==1.1.0
gunicorn==20.1.0